Replace (echo "aborted") with (assert true)
Different solvers have different handling for string literals in (echo)
commands.  The standard says that they should output the string literal
including quotes and escape characters, but most don't.

Boolector, notably, outputs the literal with quotes but does not output
a character after the closing `"`.  This causes Many-SMT to hang, since
that closing quote is ambiguous: it could be the end of a string or the
start of the two-character "" escape sequence, depending on the next
character in Boolector's output (which will never come).